dashboard/serializers.py
from .models import Device
from rest_framework import serializers


class DeviceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Device
        fields = ('pk', 'name', 'host', 'device_type', 'platform')
