requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
djangorestframework==3.9.2
future==0.17.1
idna==2.8
Jinja2==2.10
junos-eznc==2.2.0
lxml==4.3.2
MarkupSafe==1.1.1
napalm==2.4.0
ncclient==0.6.3
netaddr==0.7.19
netmiko==2.3.3
nxapi-plumbing==0.5.2
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
pyeapi==0.8.2
pyIOSXR==0.53
PyNaCl==1.3.0
pyserial==3.4
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scp==0.13.1
selectors2==2.0.1
six==1.12.0
textfsm==0.4.1
urllib3==1.24.1
